feat(citations): move westlaw citation merger to citations app
Load csv file, example csv file
